Update build script for publishing to master from CircleCI
This script was provided to us by Shiftlab. I modified it so we can use
our bots in CircleCI to actually do the publishing.

If we are ok with this script, I can do a manual run of it before turning
on CircleCI to make sure it pushes the right stuff to the right place and
the site works.