Fix Redirects and Move PyTorch Versions to Getting Started (#83)
While working on moving PyTorch versions to a tab in Getting Started, and setting up a redirect from the old page, I noticed that the docs redirect (with `redirect_url` was clashing with the default redirects using `redirect_from` - if you look at the prod site, you will notice that the latter redirects do not work any longer). So, I changed the layout name of the redirect for docs to `docs_redirect.html` and then the others use the default redirect mechanism provided by the redirect plugin.