Fix previous version conda install command
An issues was recently reported in the conda project:

- https://github.com/conda/conda/issues/12500

regarding what appears to be a typo with this project's documentation. This pull request is an attempt to resolve this. I believe you all meant to specify an extra "nvidia" channel and not to install it to an environment named "vidia", correct?

Thanks for considering this pull request ✌️